api fix + code clean up

+++ example/spring-api/src/main/java/com/example/lapisblue/lapisblue/domain/Review.java
@@ -23,7 +23,7 @@ public class Review {
     private Sales sales;
 
     @Column(name = "rating", nullable = false)
-    private Integer rating; // -1, 0, 1
+    private Float rating;
 
     @Column(name = "review_comment", columnDefinition = "TEXT")
     private String reviewComment;
